fix(scene): let heavy 3D wallpapers finish a frame and keep drawing
Three workshop scenes failed in ways that all looked like "3D does not
render."

The apply path gave up at 20 seconds. A cold MoltenVK compile of
3588579284 takes about 25 seconds, and the pipeline cache does not
remove that first wait, so the deadline is 90 seconds. The cache is
saved beside the scene shader cache so the next launch is the short one.

Perspective draws collapsed to a column or never appeared. A mat4
written into the std140 mat3 g_NormalModelMatrix spilled into the
view-projection uniform; writes are clamped to the reflected member and
the constant is 12 floats. MoltenVK still treats counter-clockwise as
front after the negative viewport, so clockwise culling removed
Saturn's sky and rings. A skybox scaled to exactly far was clipped when
the camera was not on its center; the far plane is padded by a tenth of
a percent. mdlv23 meshes above 65535 vertices store uint32 indices, and
the vertex cap is 4 million so Live Solar System's asteroid mesh is
kept.

Two covers on workshop 3662790108 were the wallpaper's own layers,
drawn wrong. An image alpha update script never reached g_Alpha, so the
intro card stayed opaque for the whole time it was applied. A text
readout that reuses the sun group's name registered second and took
getLayer("s"), so the simulation scaled that label into full-height
white bars. The label now keeps __we_text_<id> and the earlier layer
keeps the name.

View mode 3 still places most bodies smaller than a pixel, and the
intro card still covers the scene for its own first sixteen seconds
when that setting is on.

+++ upstream/renderer/external/open-wallpaper-engine/src/Scene/Parse/WPMdlParser.hpp
@@ -47,7 +47,10 @@ struct WPMdl {
         std::vector<std::array<float, 4>>    blend_weights;
         std::vector<std::array<float, 2>>    texcoords;
         std::vector<std::array<float, 2>>    texcoord2;
-        std::vector<std::array<uint16_t, 3>> indices;
+        // 32-bit once a mesh has more vertices than a uint16 index can name.
+        // High-poly bodies are stored that way; reading them as uint16 triples
+        // doubles the triangle count and the mesh is then rejected.
+        std::vector<std::array<uint32_t, 3>> indices;
 
         struct Part {
             uint32_t id { 0 };
@@ -98,7 +101,7 @@ struct WPMdl {
         std::array<float, 2>    texcoord;
     };
     std::vector<Vertex>                  vertexs;
-    std::vector<std::array<uint16_t, 3>> indices;
+    std::vector<std::array<uint32_t, 3>> indices;
 
     // std::vector<Eigen::Matrix<float, 3, 4>> bones;
     std::shared_ptr<WPPuppet> puppet;
